requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.21.1
sqlalchemy==2.0.45
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==1.10.9
pytest==7.4.0
httpx==0.24.1
